+++ Bumbodium.Data/DBModels/BranchEmployee.cs
@@ -4,21 +4,14 @@ using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Text;
 
-namespace Bumbodium.Data
+namespace Bumbodium.Data.DBModels
 {
-    public class Account
+    public class BranchEmployee
     {
-        [Key]
         public int EmployeeId { get; set; }
+        public int FiliaalId { get; set; }
 
-        [Required]
+        public Branch Filiaal { get; set; }
         public Employee Employee { get; set; }
-
-        [StringLength(64)]
-        public string Username { get; set; }
-
-        [Required]
-        [StringLength(64)]
-        public string Password { get; set; }
     }
 }
