+++ Bumbodium.Data/DBModels/DepartmentEmployee.cs
@@ -4,23 +4,24 @@ using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Text;
 
-namespace Bumbodium.Data
+namespace Bumbodium.Data.DBModels
 {
-    public class Shift
+    public class DepartmentEmployee
     {
-        [Key]
-        public int ShiftId { get; set; }
         public int EmployeeId { get; set; }
         public DepartmentType DepartmentId { get; set; }
-        
+
         public Department Department { get; set; }
         public Employee Employee { get; set; }
 
-        [Required]
-        public DateTime ShiftStartDateTime { get; set; }
-        [Required]
-        public DateTime ShiftEndDateTime { get; set; }
+        public WorkFunction WorkFunction { get; set; }
+    }
 
-
+    public enum WorkFunction
+    {
+        Fresh,
+        stockClerk,
+        TeamLeader,
+        Butcher
     }
 }
