+++ Bumbodium.Data/DBModels/Employee.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Text;
 
-namespace Bumbodium.Data
+namespace Bumbodium.Data.DBModels
 {
     public class Employee
     {
