+++ Bumbodium.Data/DBModels/Forecast.cs
