+++ Bumbodium.Data/DBModels/Presence.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Text;
 
-namespace Bumbodium.Data
+namespace Bumbodium.Data.DBModels
 {
     public class Presence
     {
