+++ Bumbodium.Data/DBModels/Shift.cs
@@ -4,24 +4,23 @@ using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Text;
 
-namespace Bumbodium.Data
+namespace Bumbodium.Data.DBModels
 {
-    public class DepartmentEmployee
+    public class Shift
     {
+        [Key]
+        public int ShiftId { get; set; }
         public int EmployeeId { get; set; }
         public DepartmentType DepartmentId { get; set; }
 
         public Department Department { get; set; }
         public Employee Employee { get; set; }
 
-        public WorkFunction WorkFunction { get; set; }
-    }
+        [Required]
+        public DateTime ShiftStartDateTime { get; set; }
+        [Required]
+        public DateTime ShiftEndDateTime { get; set; }
 
-    public enum WorkFunction
-    {
-        Fresh,
-        stockClerk,
-        TeamLeader,
-        Butcher
+
     }
 }
