+++ Bumbodium/Models/AccountViewModel.cs
@@ -1,9 +1,9 @@
-﻿using Bumbodium.Data;
+﻿using Bumbodium.Data.DBModels;
 using System.ComponentModel.DataAnnotations;
 
 namespace Bumbodium.WebApp.Models
 {
-    public class Account
+    public class AccountViewModel
     {
         public int EmployeeId { get; set; }
         public Employee Employee { get; set; }
