+++ Bumbodium/Views/Home/Login.cshtml
@@ -1,4 +1,4 @@
-﻿@model Bumbodium.WebApp.Models.Account
+﻿@model Bumbodium.WebApp.Models.AccountViewModel
 
 
 <div class="container-fluid">
