+++ Bumbodium/Views/WeekSchedule/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model Bumbodium.Data.Shift[]
+﻿@model Bumbodium.Data.DBModels.Shift[]
 @{
     ViewData["Title"] = "Index";
     //var shiftList = Model.ToArray();
